Add a choice collection component and fix latest PR comment
